Make sure table exists before drop it

+++ database/migrations/2022_07_15_121337_drop_mysql_applications_table.php
@@ -1,11 +1,9 @@
 <?php
 
 use Illuminate\Database\Migrations\Migration;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class DropMysqlApplicationsTable extends Migration
-{
+return new class () extends Migration {
     /**
      * Run the migrations.
      *
@@ -13,7 +11,9 @@ class DropMysqlApplicationsTable extends Migration
      */
     public function up()
     {
-        Schema::drop('applications');
+        if (Schema::hasTable('applications')) {
+            Schema::drop('applications');
+        }
     }
 
     /**
@@ -25,4 +25,4 @@ class DropMysqlApplicationsTable extends Migration
     {
         //
     }
-}
+};
